+++ base/docs/Docs.jl
@@ -74,15 +74,15 @@ const META    = gensym(:meta)
 const METAType = IdDict{Any,Any}
 
 function meta(m::Module)
-    if !isdefined(m, META)
+    if !isdefined(m, META) || getfield(m, META) === nothing
         initmeta(m)
     end
     return getfield(m, META)::METAType
 end
 
 function initmeta(m::Module)
-    if !isdefined(m, META)
-        Core.eval(m, :(const $META = $(METAType())))
+    if !isdefined(m, META) || getfield(m, META) === nothing
+        Core.eval(m, :($META = $(METAType())))
         push!(modules, m)
     end
     nothing